转移弹幕相关功能
